requirements.txt
absl-py==0.3.0
astor==0.7.1
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
certifi==2018.8.13
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
Cython==0.29.13
decorator==4.3.0
Django==2.1
gast==0.2.0
grpcio==1.14.1
idna==2.7
image==1.5.24
inflection==0.3.1
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
kiwisolver==1.0.1
lightfm==1.15
lxml==4.2.4
Markdown==2.6.11
matplotlib==2.2.3
more-itertools==4.3.0
numpy==1.14.5
opencv-python==3.4.2.17
pandas==0.23.4
pandas-datareader==0.6.0
parso==0.3.1
pickleshare==0.7.4
Pillow==6.1.0
pluggy==0.8.0
prompt-toolkit==1.0.15
protobuf==3.6.1
py==1.7.0
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.9.3
python-dateutil==2.7.3
pytz==2018.5
Quandl==3.4.2
requests==2.19.1
requests-file==1.4.3
requests-ftp==0.3.1
scikit-learn==0.19.2
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
tesseract==0.1.3
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
